feat: refactor current workflow of running the test script
