+++ packages/core/package.json
@@ -33,6 +33,7 @@
 	"homepage": "https://github.com/flood-io/element#readme",
 	"dependencies": {
 		"@flood/element-compiler": "^1.3.0-beta.1",
+		"@flood/element-report": "^1.3.0-beta.19",
 		"@types/faker": "^4.1.9",
 		"@types/node": "12.7.4",
 		"@types/puppeteer": "^2.0.1",
