+++ packages/flood-runner/package.json
@@ -24,6 +24,7 @@
 	"dependencies": {
 		"@flood/element-core": "^1.2.3-beta.4",
 		"@flood/node-influx": "^5.0.9",
+		"@flood/element-report": "^1.3.0-beta.19",
 		"find-root": "^1.1.0",
 		"winston": "^3.2.1"
 	}
